requirements.txt
discord.py==0.9.1
requests==2.9.1
wheel==0.24.0
ws4py==0.3.4
